Refactor AddrSpec and add AddressRewriter interface
Changed AddrSpec in request_test.go and request.go to be a pointer rather than a value. Implemented a new interface, AddressRewriter, to make adjustments to the request address prior to initiating a dialing request. An error handling case was additionally added to the Dial function in socks5_test.go. Tests were modified to reflect these changes.