feat: Add authentication support with hashed credentials (#8)
* Add authentication support with hashed credentials

This update adds support for authenticating users with hashed credentials. It introduces the `CREDENTIALS` environment variable to accommodate this feature. The `bcrypt` Golang library is used to compare hashed passwords, ensuring secure authentication. Instructions for generating compatible hashed credentials are included in the `README.md`.

* Update test credentials to hashed values

The test credentials in the SOCKS5 and credential tests have been updated from plain text to hashed values. This is a more accurate reflection of how credentials would be stored in a real scenario, thus making the tests more reliable.